apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
redis
httpx
pytest
